Адаптив акционных товаров
Исправлен баг с переходом на страницу, добавлен сырой адаптив для страницы с товаром
